+++ tests/integration-tests/tests/networking/test_cluster_networking.py
@@ -15,7 +15,7 @@ from utils import get_compute_nodes_instance_ids
 
 
 @pytest.mark.dimensions("us-west-2", "c5.xlarge", "alinux2", "slurm")
-@pytest.mark.dimensions("eu-west-2", "c5.xlarge", "centos7", "sge")
+@pytest.mark.dimensions("eu-west-2", "c5.xlarge", "centos7", "slurm")
 @pytest.mark.usefixtures("os", "scheduler", "instance")
 def test_cluster_in_private_subnet(region, pcluster_config_reader, clusters_factory):
     # This test just creates a cluster in the private subnet and just checks that no failures occur