+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 **CHANGES**
 
+- Drop support for SGE and Torque schedulers.
 - Change format and syntax of the configuration file to be used to create the cluster, from ini to YAML.
 - Deprecate `--cluster-template`, `--extra-parameters` and `--tags` parameters for the `create` command.
 - Deprecate `--cluster-template`, `--extra-parameters`, `--reset-desired` and `--yes` parameters for the `update` command.
@@ -25,6 +26,8 @@ CHANGELOG
 - Add timestamp suffix to CloudWatch Log Group name created for the cluster.
 - Remove `pcluster-config` CLI utility.
 - Remove `amis.txt` file.
+- Remove additional EBS volume attached to the head node by default. 
+- Change NICE DCV session storage path to `/home/{UserName}`.
 
 2.x.x
 ------